re-creating 1.1.5 tag from 1.1 branch

+++ olex/langdict.h
@@ -16,7 +16,7 @@
 class TLangDict  {
   TSStrPObjList<olxstr,olxstr*, true> Records;
   olxstr CurrentLanguage, CurrentLanguageEncodingStr;
-  size_t CurrentLanguageIndex;
+  int CurrentLanguageIndex;
 protected:
   void Clear();
 public:
